Updated to .NET Core 3.1; Removed redundant code
