+++ integration-tests/Tika.Client.IntegrationTests/Tika.Client.IntegrationTests/Features/ServiceAccounts/CreateDeleteServiceAccountScenario.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Net.Http;
 using System.Threading.Tasks;
 using Tika.Client.Models;
 using Xunit;
@@ -12,8 +13,8 @@ namespace Tika.Client.IntegrationTests.Features.ServiceAccounts
 
         [Fact]
         public async Task CreateDeleteServiceAccountScenarioRecipe()
-        {
-                  Given_a_tika_client();
+        { 
+            Given_a_tika_client();
             await When_a_service_account_is_created();
             await Then_it_should_be_in_the_service_accounts_list();
             await When_a_service_account_is_deleted();
@@ -22,14 +23,16 @@ namespace Tika.Client.IntegrationTests.Features.ServiceAccounts
 
         private void Given_a_tika_client()
         {
-            var baseUri = new Uri("http://localhost:3000", System.UriKind.Absolute);
-            _tikaClient = TikaClient.FromBaseUri(baseUri);
+            var httpClient = new HttpClient();
+            var options = new TikaOptions { TIKA_API_ENDPOINT = "http://localhost:3000"};
+            httpClient.BaseAddress = new Uri(options.TIKA_API_ENDPOINT, System.UriKind.Absolute);
+            _tikaClient = new TikaClient(httpClient, options);
         }
 
         private async Task When_a_service_account_is_created()
         {
             var serviceAccountName = "test-" + Guid.NewGuid().ToString().Substring(0, 5);
-            _serviceAccount = await _tikaClient.ServiceAccounts.CreateAsync(
+            _serviceAccount = await _tikaClient.CreateServiceAccount(
                 serviceAccountName,
                 "Created by integration test, may be deleted at any time."
             );
@@ -37,7 +40,7 @@ namespace Tika.Client.IntegrationTests.Features.ServiceAccounts
 
         private async Task Then_it_should_be_in_the_service_accounts_list()
         {
-            var serviceAccounts = await _tikaClient.ServiceAccounts.GetAsync();
+            var serviceAccounts = await _tikaClient.GetServiceAccounts();
 
             Assert.Contains(serviceAccounts, s =>
                 s.Id == _serviceAccount.Id &&
@@ -48,12 +51,12 @@ namespace Tika.Client.IntegrationTests.Features.ServiceAccounts
 
         private async Task When_a_service_account_is_deleted()
         {
-            await _tikaClient.ServiceAccounts.DeleteAsync(_serviceAccount.Id);
+            await _tikaClient.DeleteServiceAccount(_serviceAccount.Id);
         }
 
         private async Task Then_it_should_not_be_in_the_service_accounts_list()
         {
-            var serviceAccounts = await _tikaClient.ServiceAccounts.GetAsync();
+            var serviceAccounts = await _tikaClient.GetServiceAccounts();
 
             Assert.DoesNotContain(serviceAccounts, s =>
                 s.Id == _serviceAccount.Id &&
