--- src/KafkaJanitor.WebApp/Infrastructure/Tika/Model/ApiKey.cs
@@ -1,8 +1,0 @@
-namespace KafkaJanitor.WebApp.Infrastructure.Tika.Model
-{
-    public class ApiKey
-    {
-        public string Key { get; set; }
-        public string Secret { get; set; }
-    }
-}
