--- src/KafkaJanitor.WebApp/Infrastructure/Tika/Model/ServiceAccount.cs
@@ -1,9 +1,0 @@
-namespace KafkaJanitor.WebApp.Infrastructure.Tika.Model
-{
-    public class ServiceAccount
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
-    }
-}
