+++ src/Tika.Client/Models/ApiKey.cs
@@ -2,16 +2,7 @@ namespace Tika.Client.Models
 {
     public class ApiKey
     {
-        public string Key { get; }
-        public string Secret { get; }
-
-        public ApiKey(
-            string key, 
-            string secret
-        )
-        {
-            Key = key;
-            Secret = secret;
-        }
+        public string Key { get; set; }
+        public string Secret { get; set; }
     }
 }
