+++ src/Tika.Client/Models/ServiceAccount.cs
@@ -2,19 +2,8 @@ namespace Tika.Client.Models
 {
     public class ServiceAccount
     {
-        public int Id { get; private set; }
-        public string Name { get; private set; }
-        public string Description { get; private set; }
-
-        public ServiceAccount(
-            int id,
-            string name,
-            string description
-        )
-        {
-            Id = id;
-            Name = name;
-            Description = description;
-        }
+        public string Id { get; set; }
+        public string Name { get; set; }
+        public string Description { get; set; }
     }
 }
